Created and implementation methods: purge() and init()

+++ src/main/resources/zajavka/database/car_dealership_ddl.sql
@@ -51,7 +51,7 @@ CREATE TABLE car_to_buy
     vin             VARCHAR(20)     NOT NULL,
     brand           VARCHAR(32)     NOT NULL,
     model           VARCHAR(32)     NOT NULL,
-    year            SMALLINT        NOT NULL,
+    year            INT        NOT NULL,
     color           VARCHAR(32),
     price           NUMERIC(19, 2)  NOT NULL,
     PRIMARY KEY (car_to_buy_id),
@@ -64,7 +64,7 @@ CREATE TABLE car_to_service
     vin    VARCHAR(20)          NOT NULL,
     brand  VARCHAR(32)          NOT NULL,
     model  VARCHAR(32)          NOT NULL,
-    year   SMALLINT             NOT NULL,
+    year   INT             NOT NULL,
     PRIMARY KEY (car_to_service_id),
     UNIQUE (vin)
 );
